requirements.txt
pytorch
geneal
quadprog